some cleanup, add another hash algo and change new method for API
