+++ gokey/cache.go
@@ -8,10 +8,12 @@ import (
 
 type Cache struct {
 	sync.RWMutex
-	pairsSet map[string]pair //contains expiration time and value of a key
+	pairsSet map[string]tuple //contains expiration time and value of a key
+
+	hashFn func([]byte) (string, error)
 }
 
-type pair struct {
+type tuple struct {
 	ttl       time.Duration
 	createdAt time.Time
 	value     []byte
@@ -20,11 +22,18 @@ type pair struct {
 var (
 	_ Operations = (*Cache)(nil)
 
-	ErrEmptyKey       = errors.New("key cannot be empty")
-	ErrNoExistKey     = errors.New("key does not exist")
-	ErrNotImplemented = errors.New("not implemented")
-	ErrExpiredKey     = errors.New("key has expired")
+	ErrEmptyKey   = errors.New("key cannot be empty")
+	ErrNoExistKey = errors.New("key does not exist")
+	ErrExpiredKey = errors.New("key has expired")
 )
+
+func newCache() *Cache {
+	return &Cache{
+		RWMutex:  sync.RWMutex{},
+		pairsSet: make(map[string]tuple),
+		hashFn:   generateMD5,
+	}
+}
 
 // Get the values of the key, if this exists in the cache
 func (c *Cache) Get(key string) ([]byte, error) {
@@ -35,7 +44,11 @@ func (c *Cache) Get(key string) ([]byte, error) {
 	c.RLock()
 	defer c.RUnlock()
 
-	keyHashed := generateMD5HashFromKey([]byte(key))
+	keyHashed, err := c.hashFn([]byte(key))
+	if err != nil {
+		return nil, err
+	}
+
 	pair, exists := c.pairsSet[keyHashed]
 
 	if !exists {
@@ -61,15 +74,18 @@ func (c *Cache) Upsert(key string, value []byte, ttl time.Duration) (bool, error
 	c.Lock()
 	defer c.Unlock()
 
-	var keyHashed = generateMD5HashFromKey([]byte(key))
+	keyHashed, err := c.hashFn([]byte(key))
+	if err != nil {
+		return false, err
+	}
 
 	if c.pairsSet == nil {
-		c.pairsSet = make(map[string]pair)
+		c.pairsSet = make(map[string]tuple)
 	}
 
 	// redis in generic command:  if (ttl == -1)
 	// golang use with functions time.Duration = -1
-	c.pairsSet[keyHashed] = pair{
+	c.pairsSet[keyHashed] = tuple{
 		ttl:       ttl,
 		createdAt: time.Now(),
 		value:     value,
@@ -86,7 +102,10 @@ func (c *Cache) Delete(key string) (bool, error) {
 	c.Lock()
 	defer c.Unlock()
 
-	var keyHashed = generateMD5HashFromKey([]byte(key))
+	keyHashed, err := c.hashFn([]byte(key))
+	if err != nil {
+		return false, err
+	}
 	_, exists := c.pairsSet[keyHashed]
 
 	if exists {
@@ -106,7 +125,11 @@ func (c *Cache) Exists(key string) (bool, error) {
 	c.RLock()
 	defer c.RUnlock()
 
-	keyHashed := generateMD5HashFromKey([]byte(key))
+	keyHashed, err := c.hashFn([]byte(key))
+	if err != nil {
+		return false, err
+	}
+
 	pair, exists := c.pairsSet[keyHashed]
 
 	if !exists {
